+++ lib/src/entities/user.dart
@@ -4,8 +4,13 @@ abstract class User extends RestrrEntity {
   int get id;
   String get username;
   String? get email;
+  String? get displayName;
   DateTime get createdAt;
   bool get isAdmin;
+
+  /// Returns the effective display name for the user.
+  /// If the display name is not set, the username is returned instead.
+  String get effectiveDisplayName;
 }
 
 class UserImpl extends RestrrEntityImpl implements User {
@@ -16,6 +21,8 @@ class UserImpl extends RestrrEntityImpl implements User {
   @override
   final String? email;
   @override
+  final String? displayName;
+  @override
   final DateTime createdAt;
   @override
   final bool isAdmin;
@@ -25,7 +32,11 @@ class UserImpl extends RestrrEntityImpl implements User {
     required this.id,
     required this.username,
     required this.email,
+    required this.displayName,
     required this.createdAt,
     required this.isAdmin,
   });
+
+  @override
+  String get effectiveDisplayName => displayName ?? username;
 }